+++ mesheryctl/internal/cli/root/system/fixtures/.meshery/meshery.yaml
@@ -1,0 +1,30 @@
+services:
+  meshery:
+    image: layer5/meshery:stable-v0.5.10
+    labels:
+    - com.centurylinklabs.watchtower.enable=true
+    environment:
+    - PROVIDER_BASE_URLS=https://meshery.layer5.io
+    - SAAS_BASE_URL=https://meshery.layer5.io
+    - ADAPTER_URLS=meshery-istio:10000 meshery-linkerd:10001 meshery-consul:10002
+      meshery-octarine:10003 meshery-nsm:10004 meshery-kuma:10007 meshery-cpx:10008
+      meshery-osm:10009 meshery-traefik-mesh:10006
+    - EVENT=mesheryLocal
+    - MESHERY_SERVER_CALLBACK_URL=
+    volumes:
+    - $HOME/.kube:/home/appuser/.kube:ro
+    ports:
+    - 9081:8080
+  meshery-istio:
+    image: layer5/meshery-istio:stable-latest
+    labels:
+    - com.centurylinklabs.watchtower.enable=true
+    ports:
+    - 10000:10000
+  watchtower:
+    image: v2tec/watchtower
+    labels:
+    - com.centurylinklabs.watchtower.enable=true
+    volumes:
+    - /var/run/docker.sock:/var/run/docker.sock
+version: "3"
