+++ mesheryctl/internal/cli/root/system/testdata/reset/reset_kubernetes.output.golden
@@ -5,6 +5,6 @@ Channel: stable
 Version: v0.5.40
 Platform: kubernetes
 
-Fetching Meshery Server and Meshery Operator manifests...
+Fetching Meshery Server and Meshery Operator manifests for  local context...
 downloading manifest files from https://github.com/meshery/meshery/tree/v0.5.40/install/deployment_yamls/k8s
 ...meshconfig has been reset to default settings.
